refactor: split test_db_query.py and standardize all test fixtures
Split test_db_query.py (3200 lines) into 9 category-based files and
apply fixture-based DB management across ALL test files:

Phase 1 — Split test_db_query.py:
- test_schema.py, test_dml.py, test_query.py, test_string_functions.py
- test_path.py, test_intersect.py, test_optional_match.py
- test_persistence.py, test_service.py

Phase 2 — Standardize all remaining test files:
- test_db_transaction.py: /tmp/modern_graph → modern_graph fixture
- test_ngcli_basics.py, test_ngcli_commands.py: same
- test_export.py, test_load.py: /tmp/tinysnb → tinysnb fixture,
  /tmp/comprehensive_graph → skipif
- test_db_connection.py, test_tp_service.py, test_batch_loading.py:
  /tmp/test_* → tmp_path
- test_db_init.py: /tmp/ldbc → skipif + LDBC_DIR

Fixtures in conftest.py:
- modern_graph, tinysnb, empty_db (tmp_path-isolated)
- HAS_LDBC, HAS_COMPREHENSIVE_GRAPH (skipif flags)

All fixtures use tmp_path (unique per test), eliminating:
- Race conditions in parallel test execution
- Cross-test data pollution from shared /tmp/* paths

Other fixes:
- Aligned graph.yaml with load_builtin_dataset schema
- Updated CI workflow to run new test files
- Format-checked all files (isort + black + flake8)

Closes #609.
