+++ example_dataset/modern_graph/graph.yaml
@@ -74,3 +74,7 @@ schema:
           property_name: weight
           property_type:
             primitive_type: DT_DOUBLE
+        - property_id: 1
+          property_name: since
+          property_type:
+            primitive_type: DT_SIGNED_INT32
